\name{MaizeNAM-package}
\alias{MaizeNAM-package}
\docType{package}
\title{
Maize Nested Association Mapping Dataset
}
\description{
    Genomic and multi-environmental maize data. Maize Nested
    Association Mapping (NAM) project dataset funded by the National Science Foundation.
    BLUP function formats data for genome-wide prediction and association analysis.
}
\details{
\tabular{ll}{
Package: \tab MaizeNAM\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2021-04-11\cr
License: \tab GPL-3 \cr
}
}
\author{
Brian Rice, Alencar Xavier, Alex Lipka
Maintainer: Brian Rice <ricebrian06@gmail.com>
}
\seealso{
Functions: BLUP, data(pheno), data(geno) and data(bins).
}
